feat: initial bot code (#67)
* feat: establish initial bot

Sets up the initial bot infrastructure. Includes a basic help and
ping command for testing purposes. Builds out the interfaces needed
to allow additional commands.

* fix: sort commands in help

Adds a fix to send the command list in the help command in order.

* 0.1.0

* fix: interface naming conventions

Modifies the naming conventions for the interfaces.

* docs: how to create commands

Adds brief documentation on the process for creating commands.

Co-authored-by: Eddie Jaoude <eddie@jaoudestudios.com>